Fixed Issues With Types
Prevented "possible data loss" by replacing some ints with size_t

+++ src/allocator/allocator.cpp
@@ -5,7 +5,7 @@ namespace lithe {
     allocator::allocator(
         const lithe::buffer& buff_,
         const std::vector<size_t>& sizes_,
-        const std::vector<int>& starting_
+        const std::vector<size_t>& starting_
     ):
         buff(buff_),
         sizes(sizes_),
